Chgt pour les potions pour être + orienté objet
