+++ src/model/CS.java
@@ -33,7 +33,7 @@ public class CS extends Player {
 		//Add weapons and items to the Player at the beginning of the Game.
 		try {
 			inventory.addWeapon(new Sword(0,0,this.getGame()));
-			inventory.addConsumable(new Potion(Potion.potionType.vie, 0, 0, this.getGame())); //Note the type of potion
+			inventory.addConsumable(new PotionVie(0, 0, this.getGame())); //Note the type of potion
 			inventory.addSpell(new Fire(0,0,game,true));
 			inventory.addSpell(new Thunder(0,0,game,true));
 			inventory.addSpell(new Ice(0,0,game,true));
